requirements.txt
firebase-admin==6.4.0
aiohttp==3.9.3
websockets==10.4
jsonalias==0.1.1
solana==0.27.0
solders==0.9.3
base58==2.1.1
loguru==0.7.2
httpx==0.23.3
python-dotenv==1.0.0
requests==2.31.0
construct==2.10.68
construct-typing==0.5.6
cachetools==4.2.2
anyio==3.7.1
multidict==6.0.4
yarl==1.9.4
frozenlist==1.4.1
attrs==23.2.0
aiosignal==1.3.1
idna==3.6
charset-normalizer==3.3.2
async-timeout==4.0.3
h11==0.14.0
rfc3986==1.5.0
sniffio==1.3.0
certifi==2023.11.17
typing-extensions==4.9.0
urllib3==2.1.0
cryptography==41.0.7
pytz==2023.3.post1
aiogram>=3.3.0